+++ exporter/opencensusexporter/factory_test.go
@@ -24,7 +24,6 @@ import (
 	"go.uber.org/zap"
 
 	"go.opentelemetry.io/collector/component"
-	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configcheck"
 	"go.opentelemetry.io/collector/config/configgrpc"
@@ -41,10 +40,9 @@ func TestCreateDefaultConfig(t *testing.T) {
 func TestCreateTracesExporter(t *testing.T) {
 	endpoint := testutil.GetAvailableLocalAddress(t)
 	tests := []struct {
-		name            string
-		config          Config
-		mustFail        bool
-		mustFailOnStart bool
+		name     string
+		config   Config
+		mustFail bool
 	}{
 		{
 			name: "NoEndpoint",
@@ -134,8 +132,7 @@ func TestCreateTracesExporter(t *testing.T) {
 				},
 				NumWorkers: 3,
 			},
-			mustFail:        false,
-			mustFailOnStart: true,
+			mustFail: true,
 		},
 		{
 			name: "CaCert",
@@ -166,35 +163,28 @@ func TestCreateTracesExporter(t *testing.T) {
 				},
 				NumWorkers: 3,
 			},
-			mustFail:        false,
-			mustFailOnStart: true,
+			mustFail: true,
 		},
 	}
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			params := component.ExporterCreateParams{Logger: zap.NewNop()}
-			tExporter, tErr := createTracesExporter(context.Background(), params, &tt.config)
-			checkErrorsAndStartAndShutdown(t, tExporter, tErr, tt.mustFail, tt.mustFailOnStart)
-			mExporter, mErr := createMetricsExporter(context.Background(), params, &tt.config)
-			checkErrorsAndStartAndShutdown(t, mExporter, mErr, tt.mustFail, tt.mustFailOnStart)
+			tReceiver, tErr := createTracesExporter(context.Background(), params, &tt.config)
+			checkErrorsAndShutdown(t, tReceiver, tErr, tt.mustFail)
+			mReceiver, mErr := createMetricsExporter(context.Background(), params, &tt.config)
+			checkErrorsAndShutdown(t, mReceiver, mErr, tt.mustFail)
 		})
 	}
 }
 
-func checkErrorsAndStartAndShutdown(t *testing.T, exporter component.Exporter, err error, mustFail, mustFailOnStart bool) {
+func checkErrorsAndShutdown(t *testing.T, receiver component.Receiver, err error, mustFail bool) {
 	if mustFail {
 		assert.NotNil(t, err)
-		return
-	}
-	assert.NoError(t, err)
-	assert.NotNil(t, exporter)
+	} else {
+		assert.NoError(t, err)
+		assert.NotNil(t, receiver)
 
-	sErr := exporter.Start(context.Background(), componenttest.NewNopHost())
-	if mustFailOnStart {
-		require.Error(t, sErr)
-		return
+		require.NoError(t, receiver.Shutdown(context.Background()))
 	}
-	require.NoError(t, sErr)
-	require.NoError(t, exporter.Shutdown(context.Background()))
 }