+++ exporter/prometheusremotewriteexporter/exporter_test.go
@@ -33,13 +33,14 @@ import (
 	"go.opentelemetry.io/collector/component/componenttest"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/confighttp"
+	"go.opentelemetry.io/collector/config/configtls"
 	"go.opentelemetry.io/collector/consumer/pdata"
 	"go.opentelemetry.io/collector/exporter/exporterhelper"
 	"go.opentelemetry.io/collector/internal/testdata"
 )
 
-// Test_ NewPrwExporter checks that a new exporter instance with non-nil fields is initialized
-func Test_NewPrwExporter(t *testing.T) {
+// Test_NewPRWExporter checks that a new exporter instance with non-nil fields is initialized
+func Test_NewPRWExporter(t *testing.T) {
 	cfg := &Config{
 		ExporterSettings:   config.NewExporterSettings(config.NewID(typeStr)),
 		TimeoutSettings:    exporterhelper.TimeoutSettings{},
@@ -54,59 +55,52 @@ func Test_NewPrwExporter(t *testing.T) {
 	}
 
 	tests := []struct {
-		name           string
-		config         *Config
-		namespace      string
-		endpoint       string
-		concurrency    int
-		externalLabels map[string]string
-		client         *http.Client
-		returnError    bool
-		buildInfo      component.BuildInfo
+		name                string
+		config              *Config
+		namespace           string
+		endpoint            string
+		concurrency         int
+		externalLabels      map[string]string
+		returnErrorOnCreate bool
+		buildInfo           component.BuildInfo
 	}{
 		{
-			"invalid_URL",
-			cfg,
-			"test",
-			"invalid URL",
-			5,
-			map[string]string{"Key1": "Val1"},
-			http.DefaultClient,
-			true,
-			buildInfo,
+			name:                "invalid_URL",
+			config:              cfg,
+			namespace:           "test",
+			endpoint:            "invalid URL",
+			concurrency:         5,
+			externalLabels:      map[string]string{"Key1": "Val1"},
+			returnErrorOnCreate: true,
+			buildInfo:           buildInfo,
 		},
 		{
-			"invalid_labels_case",
-			cfg,
-			"test",
-			"http://some.url:9411/api/prom/push",
-			5,
-			map[string]string{"Key1": ""},
-			http.DefaultClient,
-			true,
-			buildInfo,
+			name:                "invalid_labels_case",
+			config:              cfg,
+			namespace:           "test",
+			endpoint:            "http://some.url:9411/api/prom/push",
+			concurrency:         5,
+			externalLabels:      map[string]string{"Key1": ""},
+			returnErrorOnCreate: true,
+			buildInfo:           buildInfo,
 		},
 		{
-			"success_case",
-			cfg,
-			"test",
-			"http://some.url:9411/api/prom/push",
-			5,
-			map[string]string{"Key1": "Val1"},
-			http.DefaultClient,
-			false,
-			buildInfo,
+			name:           "success_case",
+			config:         cfg,
+			namespace:      "test",
+			endpoint:       "http://some.url:9411/api/prom/push",
+			concurrency:    5,
+			externalLabels: map[string]string{"Key1": "Val1"},
+			buildInfo:      buildInfo,
 		},
 		{
-			"success_case_no_labels",
-			cfg,
-			"test",
-			"http://some.url:9411/api/prom/push",
-			5,
-			map[string]string{},
-			http.DefaultClient,
-			false,
-			buildInfo,
+			name:           "success_case_no_labels",
+			config:         cfg,
+			namespace:      "test",
+			endpoint:       "http://some.url:9411/api/prom/push",
+			concurrency:    5,
+			externalLabels: map[string]string{},
+			buildInfo:      buildInfo,
 		},
 	}
 
@@ -116,28 +110,104 @@ func Test_NewPrwExporter(t *testing.T) {
 			cfg.ExternalLabels = tt.externalLabels
 			cfg.Namespace = tt.namespace
 			cfg.RemoteWriteQueue.NumConsumers = 1
-			prwe, err := NewPrwExporter(cfg, tt.buildInfo)
+			prwe, err := NewPRWExporter(cfg, tt.buildInfo)
 
-			if tt.returnError {
+			if tt.returnErrorOnCreate {
 				assert.Error(t, err)
 				return
 			}
-			require.NoError(t, err)
-			require.NoError(t, prwe.Start(context.Background(), componenttest.NewNopHost()))
+			require.NotNil(t, prwe)
 			assert.NotNil(t, prwe.namespace)
 			assert.NotNil(t, prwe.endpointURL)
 			assert.NotNil(t, prwe.externalLabels)
-			assert.NotNil(t, prwe.client)
 			assert.NotNil(t, prwe.closeChan)
 			assert.NotNil(t, prwe.wg)
 			assert.NotNil(t, prwe.userAgentHeader)
+			assert.NotNil(t, prwe.clientSettings)
+		})
+	}
+}
+
+// Test_Start checks if the client is properly created as expected.
+func Test_Start(t *testing.T) {
+	cfg := &Config{
+		ExporterSettings: config.NewExporterSettings(config.NewID(typeStr)),
+		TimeoutSettings:  exporterhelper.TimeoutSettings{},
+		RetrySettings:    exporterhelper.RetrySettings{},
+		Namespace:        "",
+		ExternalLabels:   map[string]string{},
+	}
+	buildInfo := component.BuildInfo{
+		Description: "OpenTelemetry Collector",
+		Version:     "1.0",
+	}
+	tests := []struct {
+		name                 string
+		config               *Config
+		namespace            string
+		concurrency          int
+		externalLabels       map[string]string
+		returnErrorOnStartUp bool
+		buildInfo            component.BuildInfo
+		endpoint             string
+		clientSettings       confighttp.HTTPClientSettings
+	}{
+		{
+			name:           "success_case",
+			config:         cfg,
+			namespace:      "test",
+			concurrency:    5,
+			externalLabels: map[string]string{"Key1": "Val1"},
+			buildInfo:      buildInfo,
+			clientSettings: confighttp.HTTPClientSettings{Endpoint: "https://some.url:9411/api/prom/push"},
+		},
+		{
+			name:                 "invalid_tls",
+			config:               cfg,
+			namespace:            "test",
+			concurrency:          5,
+			externalLabels:       map[string]string{"Key1": "Val1"},
+			buildInfo:            buildInfo,
+			returnErrorOnStartUp: true,
+			clientSettings: confighttp.HTTPClientSettings{
+				Endpoint: "https://some.url:9411/api/prom/push",
+				TLSSetting: configtls.TLSClientSetting{
+					TLSSetting: configtls.TLSSetting{
+						CAFile:   "non-existent file",
+						CertFile: "",
+						KeyFile:  "",
+					},
+					Insecure:   false,
+					ServerName: "",
+				},
+			},
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			cfg.ExternalLabels = tt.externalLabels
+			cfg.Namespace = tt.namespace
+			cfg.RemoteWriteQueue.NumConsumers = 1
+			cfg.HTTPClientSettings = tt.clientSettings
+
+			prwe, err := NewPRWExporter(cfg, tt.buildInfo)
+			assert.NoError(t, err)
+			assert.NotNil(t, prwe)
+
+			err = prwe.Start(context.Background(), componenttest.NewNopHost())
+			if tt.returnErrorOnStartUp {
+				assert.Error(t, err)
+				return
+			}
+			assert.NotNil(t, prwe.client)
 		})
 	}
 }
 
 // Test_Shutdown checks after Shutdown is called, incoming calls to PushMetrics return error.
 func Test_Shutdown(t *testing.T) {
-	prwe := &PrwExporter{
+	prwe := &PRWExporter{
 		wg:        new(sync.WaitGroup),
 		closeChan: make(chan struct{}),
 	}
@@ -197,11 +267,11 @@ func Test_export(t *testing.T) {
 	// Create in test table format to check if different HTTP response codes or server errors
 	// are properly identified
 	tests := []struct {
-		name             string
-		ts               prompb.TimeSeries
-		serverUp         bool
-		httpResponseCode int
-		returnError      bool
+		name                string
+		ts                  prompb.TimeSeries
+		serverUp            bool
+		httpResponseCode    int
+		returnErrorOnCreate bool
 	}{
 		{"success_case",
 			*ts1,
@@ -238,7 +308,7 @@ func Test_export(t *testing.T) {
 				server.Close()
 			}
 			errs := runExportPipeline(ts1, serverURL)
-			if tt.returnError {
+			if tt.returnErrorOnCreate {
 				assert.Error(t, errs[0])
 				return
 			}
@@ -263,14 +333,13 @@ func runExportPipeline(ts *prompb.TimeSeries, endpoint *url.URL) []error {
 		Version:     "1.0",
 	}
 	// after this, instantiate a CortexExporter with the current HTTP client and endpoint set to passed in endpoint
-	prwe, err := NewPrwExporter(cfg, buildInfo)
+	prwe, err := NewPRWExporter(cfg, buildInfo)
 	if err != nil {
 		errs = append(errs, err)
 		return errs
 	}
 
-	err = prwe.Start(context.Background(), componenttest.NewNopHost())
-	if err != nil {
+	if err = prwe.Start(context.Background(), componenttest.NewNopHost()); err != nil {
 		errs = append(errs, err)
 		return errs
 	}
@@ -515,16 +584,16 @@ func Test_PushMetrics(t *testing.T) {
 				},
 				RemoteWriteQueue: RemoteWriteQueue{NumConsumers: 5},
 			}
-
 			assert.NotNil(t, cfg)
+			// c, err := config.HTTPClientSettings.ToClient()
+			// assert.Nil(t, err)
 			buildInfo := component.BuildInfo{
 				Description: "OpenTelemetry Collector",
 				Version:     "1.0",
 			}
-			prwe, nErr := NewPrwExporter(cfg, buildInfo)
+			prwe, nErr := NewPRWExporter(cfg, buildInfo)
 			require.NoError(t, nErr)
 			require.NoError(t, prwe.Start(context.Background(), componenttest.NewNopHost()))
-
 			err := prwe.PushMetrics(context.Background(), *tt.md)
 			if tt.returnErr {
 				assert.Error(t, err)
@@ -537,10 +606,10 @@ func Test_PushMetrics(t *testing.T) {
 
 func Test_validateAndSanitizeExternalLabels(t *testing.T) {
 	tests := []struct {
-		name           string
-		inputLabels    map[string]string
-		expectedLabels map[string]string
-		returnError    bool
+		name                string
+		inputLabels         map[string]string
+		expectedLabels      map[string]string
+		returnErrorOnCreate bool
 	}{
 		{"success_case_no_labels",
 			map[string]string{},
@@ -572,7 +641,7 @@ func Test_validateAndSanitizeExternalLabels(t *testing.T) {
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
 			newLabels, err := validateAndSanitizeExternalLabels(tt.inputLabels)
-			if tt.returnError {
+			if tt.returnErrorOnCreate {
 				assert.Error(t, err)
 				return
 			}