+++ exporter/zipkinexporter/zipkin.go
@@ -65,13 +65,9 @@ func createZipkinExporter(cfg *Config) (*zipkinExporter, error) {
 }
 
 // start creates the http client
-func (ze *zipkinExporter) start(_ context.Context, host component.Host) error {
-	client, err := ze.clientSettings.ToClient(host.GetExtensions())
-	if err != nil {
-		return err
-	}
-	ze.client = client
-	return nil
+func (ze *zipkinExporter) start(_ context.Context, _ component.Host) (err error) {
+	ze.client, err = ze.clientSettings.ToClient()
+	return
 }
 
 func (ze *zipkinExporter) pushTraceData(ctx context.Context, td pdata.Traces) error {