+++ obsreport/obsreport_receiver.go
@@ -23,72 +23,8 @@ import (
 
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configtelemetry"
-)
-
-const (
-	// ReceiverKey used to identify receivers in metrics and traces.
-	ReceiverKey = "receiver"
-	// TransportKey used to identify the transport used to received the data.
-	TransportKey = "transport"
-	// FormatKey used to identify the format of the data received.
-	FormatKey = "format"
-
-	// AcceptedSpansKey used to identify spans accepted by the Collector.
-	AcceptedSpansKey = "accepted_spans"
-	// RefusedSpansKey used to identify spans refused (ie.: not ingested) by the Collector.
-	RefusedSpansKey = "refused_spans"
-
-	// AcceptedMetricPointsKey used to identify metric points accepted by the Collector.
-	AcceptedMetricPointsKey = "accepted_metric_points"
-	// RefusedMetricPointsKey used to identify metric points refused (ie.: not ingested) by the
-	// Collector.
-	RefusedMetricPointsKey = "refused_metric_points"
-
-	// AcceptedLogRecordsKey used to identify log records accepted by the Collector.
-	AcceptedLogRecordsKey = "accepted_log_records"
-	// RefusedLogRecordsKey used to identify log records refused (ie.: not ingested) by the
-	// Collector.
-	RefusedLogRecordsKey = "refused_log_records"
-)
-
-var (
-	tagKeyReceiver, _  = tag.NewKey(ReceiverKey)
-	tagKeyTransport, _ = tag.NewKey(TransportKey)
-
-	receiverPrefix                  = ReceiverKey + nameSep
-	receiveTraceDataOperationSuffix = nameSep + "TraceDataReceived"
-	receiverMetricsOperationSuffix  = nameSep + "MetricsReceived"
-	receiverLogsOperationSuffix     = nameSep + "LogsReceived"
-
-	// Receiver metrics. Any count of data items below is in the original format
-	// that they were received, reasoning: reconciliation is easier if measurements
-	// on clients and receiver are expected to be the same. Translation issues
-	// that result in a different number of elements should be reported in a
-	// separate way.
-	mReceiverAcceptedSpans = stats.Int64(
-		receiverPrefix+AcceptedSpansKey,
-		"Number of spans successfully pushed into the pipeline.",
-		stats.UnitDimensionless)
-	mReceiverRefusedSpans = stats.Int64(
-		receiverPrefix+RefusedSpansKey,
-		"Number of spans that could not be pushed into the pipeline.",
-		stats.UnitDimensionless)
-	mReceiverAcceptedMetricPoints = stats.Int64(
-		receiverPrefix+AcceptedMetricPointsKey,
-		"Number of metric points successfully pushed into the pipeline.",
-		stats.UnitDimensionless)
-	mReceiverRefusedMetricPoints = stats.Int64(
-		receiverPrefix+RefusedMetricPointsKey,
-		"Number of metric points that could not be pushed into the pipeline.",
-		stats.UnitDimensionless)
-	mReceiverAcceptedLogRecords = stats.Int64(
-		receiverPrefix+AcceptedLogRecordsKey,
-		"Number of log records successfully pushed into the pipeline.",
-		stats.UnitDimensionless)
-	mReceiverRefusedLogRecords = stats.Int64(
-		receiverPrefix+RefusedLogRecordsKey,
-		"Number of log records that could not be pushed into the pipeline.",
-		stats.UnitDimensionless)
+	"go.opentelemetry.io/collector/internal/obsreportconfig"
+	"go.opentelemetry.io/collector/internal/obsreportconfig/obsmetrics"
 )
 
 // StartReceiveOptions has the options related to starting a receive operation.
@@ -172,23 +108,7 @@ func (rec *Receiver) StartTraceDataReceiveOp(
 ) context.Context {
 	return rec.traceReceiveOp(
 		operationCtx,
-		receiveTraceDataOperationSuffix,
-		opt...)
-}
-
-// StartTraceDataReceiveOp is deprecated but is called when a request is received from a client.
-// The returned context should be used in other calls to the obsreport functions
-// dealing with the same receive operation.
-func StartTraceDataReceiveOp(
-	operationCtx context.Context,
-	receiverID config.ComponentID,
-	transport string,
-	opt ...StartReceiveOption,
-) context.Context {
-	rec := NewReceiver(ReceiverSettings{ReceiverID: receiverID, Transport: transport})
-	return rec.traceReceiveOp(
-		operationCtx,
-		receiveTraceDataOperationSuffix,
+		obsmetrics.ReceiveTraceDataOperationSuffix,
 		opt...)
 }
 
@@ -209,24 +129,6 @@ func (rec *Receiver) EndTraceDataReceiveOp(
 	)
 }
 
-// EndTraceDataReceiveOp is deprecated but completes the receive operation that was started with
-// StartTraceDataReceiveOp.
-func EndTraceDataReceiveOp(
-	receiverCtx context.Context,
-	format string,
-	numReceivedSpans int,
-	err error,
-) {
-	rec := NewReceiver(ReceiverSettings{})
-	rec.endReceiveOp(
-		receiverCtx,
-		format,
-		numReceivedSpans,
-		err,
-		config.TracesDataType,
-	)
-}
-
 // StartLogsReceiveOp is called when a request is received from a client.
 // The returned context should be used in other calls to the obsreport functions
 // dealing with the same receive operation.
@@ -236,23 +138,7 @@ func (rec *Receiver) StartLogsReceiveOp(
 ) context.Context {
 	return rec.traceReceiveOp(
 		operationCtx,
-		receiverLogsOperationSuffix,
-		opt...)
-}
-
-// StartLogsReceiveOp is deprecated but is called when a request is received from a client.
-// The returned context should be used in other calls to the obsreport functions
-// dealing with the same receive operation.
-func StartLogsReceiveOp(
-	operationCtx context.Context,
-	receiverID config.ComponentID,
-	transport string,
-	opt ...StartReceiveOption,
-) context.Context {
-	rec := NewReceiver(ReceiverSettings{ReceiverID: receiverID, Transport: transport})
-	return rec.traceReceiveOp(
-		operationCtx,
-		receiverLogsOperationSuffix,
+		obsmetrics.ReceiverLogsOperationSuffix,
 		opt...)
 }
 
@@ -273,24 +159,6 @@ func (rec *Receiver) EndLogsReceiveOp(
 	)
 }
 
-// EndLogsReceiveOp is deprecated but completes the receive operation that was started with
-// StartLogsReceiveOp.
-func EndLogsReceiveOp(
-	receiverCtx context.Context,
-	format string,
-	numReceivedLogRecords int,
-	err error,
-) {
-	rec := NewReceiver(ReceiverSettings{})
-	rec.endReceiveOp(
-		receiverCtx,
-		format,
-		numReceivedLogRecords,
-		err,
-		config.LogsDataType,
-	)
-}
-
 // StartMetricsReceiveOp is called when a request is received from a client.
 // The returned context should be used in other calls to the obsreport functions
 // dealing with the same receive operation.
@@ -300,23 +168,7 @@ func (rec *Receiver) StartMetricsReceiveOp(
 ) context.Context {
 	return rec.traceReceiveOp(
 		operationCtx,
-		receiverMetricsOperationSuffix,
-		opt...)
-}
-
-// StartMetricsReceiveOp is deprecated but is called when a request is received from a client.
-// The returned context should be used in other calls to the obsreport functions
-// dealing with the same receive operation.
-func StartMetricsReceiveOp(
-	operationCtx context.Context,
-	receiverID config.ComponentID,
-	transport string,
-	opt ...StartReceiveOption,
-) context.Context {
-	rec := NewReceiver(ReceiverSettings{ReceiverID: receiverID, Transport: transport})
-	return rec.traceReceiveOp(
-		operationCtx,
-		receiverMetricsOperationSuffix,
+		obsmetrics.ReceiverMetricsOperationSuffix,
 		opt...)
 }
 
@@ -337,24 +189,6 @@ func (rec *Receiver) EndMetricsReceiveOp(
 	)
 }
 
-// EndMetricsReceiveOp is deprecated but completes the receive operation that was started with
-// StartMetricsReceiveOp.
-func EndMetricsReceiveOp(
-	receiverCtx context.Context,
-	format string,
-	numReceivedPoints int,
-	err error,
-) {
-	rec := NewReceiver(ReceiverSettings{})
-	rec.endReceiveOp(
-		receiverCtx,
-		format,
-		numReceivedPoints,
-		err,
-		config.MetricsDataType,
-	)
-}
-
 // ReceiverContext adds the keys used when recording observability metrics to
 // the given context returning the newly created context. This context should
 // be used in related calls to the obsreport functions so metrics are properly
@@ -365,8 +199,8 @@ func ReceiverContext(
 	transport string,
 ) context.Context {
 	ctx, _ = tag.New(ctx,
-		tag.Upsert(tagKeyReceiver, receiverID.String(), tag.WithTTL(tag.TTLNoPropagation)),
-		tag.Upsert(tagKeyTransport, transport, tag.WithTTL(tag.TTLNoPropagation)))
+		tag.Upsert(obsmetrics.TagKeyReceiver, receiverID.String(), tag.WithTTL(tag.TTLNoPropagation)),
+		tag.Upsert(obsmetrics.TagKeyTransport, transport, tag.WithTTL(tag.TTLNoPropagation)))
 
 	return ctx
 }
@@ -385,7 +219,7 @@ func (rec *Receiver) traceReceiveOp(
 
 	var ctx context.Context
 	var span *trace.Span
-	spanName := receiverPrefix + rec.receiverID.String() + operationSuffix
+	spanName := obsmetrics.ReceiverPrefix + rec.receiverID.String() + operationSuffix
 	if !opts.LongLivedCtx {
 		ctx, span = trace.StartSpan(receiverCtx, spanName)
 	} else {
@@ -401,7 +235,7 @@ func (rec *Receiver) traceReceiveOp(
 	}
 
 	if rec.transport != "" {
-		span.AddAttributes(trace.StringAttribute(TransportKey, rec.transport))
+		span.AddAttributes(trace.StringAttribute(obsmetrics.TransportKey, rec.transport))
 	}
 	return ctx
 }
@@ -423,18 +257,18 @@ func (rec *Receiver) endReceiveOp(
 
 	span := trace.FromContext(receiverCtx)
 
-	if gLevel != configtelemetry.LevelNone {
+	if obsreportconfig.Level != configtelemetry.LevelNone {
 		var acceptedMeasure, refusedMeasure *stats.Int64Measure
 		switch dataType {
 		case config.TracesDataType:
-			acceptedMeasure = mReceiverAcceptedSpans
-			refusedMeasure = mReceiverRefusedSpans
+			acceptedMeasure = obsmetrics.ReceiverAcceptedSpans
+			refusedMeasure = obsmetrics.ReceiverRefusedSpans
 		case config.MetricsDataType:
-			acceptedMeasure = mReceiverAcceptedMetricPoints
-			refusedMeasure = mReceiverRefusedMetricPoints
+			acceptedMeasure = obsmetrics.ReceiverAcceptedMetricPoints
+			refusedMeasure = obsmetrics.ReceiverRefusedMetricPoints
 		case config.LogsDataType:
-			acceptedMeasure = mReceiverAcceptedLogRecords
-			refusedMeasure = mReceiverRefusedLogRecords
+			acceptedMeasure = obsmetrics.ReceiverAcceptedLogRecords
+			refusedMeasure = obsmetrics.ReceiverRefusedLogRecords
 		}
 
 		stats.Record(
@@ -448,19 +282,19 @@ func (rec *Receiver) endReceiveOp(
 		var acceptedItemsKey, refusedItemsKey string
 		switch dataType {
 		case config.TracesDataType:
-			acceptedItemsKey = AcceptedSpansKey
-			refusedItemsKey = RefusedSpansKey
+			acceptedItemsKey = obsmetrics.AcceptedSpansKey
+			refusedItemsKey = obsmetrics.RefusedSpansKey
 		case config.MetricsDataType:
-			acceptedItemsKey = AcceptedMetricPointsKey
-			refusedItemsKey = RefusedMetricPointsKey
+			acceptedItemsKey = obsmetrics.AcceptedMetricPointsKey
+			refusedItemsKey = obsmetrics.RefusedMetricPointsKey
 		case config.LogsDataType:
-			acceptedItemsKey = AcceptedLogRecordsKey
-			refusedItemsKey = RefusedLogRecordsKey
+			acceptedItemsKey = obsmetrics.AcceptedLogRecordsKey
+			refusedItemsKey = obsmetrics.RefusedLogRecordsKey
 		}
 
 		span.AddAttributes(
 			trace.StringAttribute(
-				FormatKey, format),
+				obsmetrics.FormatKey, format),
 			trace.Int64Attribute(
 				acceptedItemsKey, int64(numAccepted)),
 			trace.Int64Attribute(